# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Motorcycle Rental & Catalog API 🏍️",
    "version": "1.0.0",
    "description": "API REST robusta en Node.js y TypeScript para la gestión de catálogo de motocicletas, reservas con validación de fechas, opiniones de usuarios y autenticación basada en JWT."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ingresa tu token JWT en formato: `Bearer <token>`"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Juan Pérez" },
          "email": { "type": "string", "format": "email", "example": "juan@example.com" },
          "role": { "type": "string", "enum": ["USER", "ADMIN"], "example": "USER" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 2 },
          "motorcycleId": { "type": "integer", "example": 1 },
          "startDate": { "type": "string", "format": "date-time", "example": "2026-08-06T00:00:00Z" },
          "endDate": { "type": "string", "format": "date-time", "example": "2026-08-10T00:00:00Z" },
          "status": { "type": "string", "enum": ["PENDING", "APPROVED", "CANCELLED"], "example": "PENDING" },
          "totalPrice": { "type": "number", "format": "float", "example": 380.0 }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar un nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Juan Pérez" },
                  "email": { "type": "string", "format": "email", "example": "juan@example.com" },
                  "password": { "type": "string", "example": "password123" },
                  "role": { "type": "string", "enum": ["USER", "ADMIN"], "default": "USER", "example": "USER" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "User registered successfully" },
                    "token": { "type": "string", "example": "eyJhbGciOiJI..." },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Datos de registro inválidos o email ya registrado" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión de usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "juan@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Login successful" },
                    "token": { "type": "string", "example": "eyJhbGciOiJI..." },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Email o contraseña incorrectos" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Obtener perfil del usuario autenticado",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Detalles del perfil obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "No autorizado (Falta token o es inválido)" }
        }
      }
    },
...
    "/api/bookings": {
      "get": {
        "tags": ["Reservas"],
        "summary": "Listar reservas (Admins ven todas, usuarios solo las suyas)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "description": "Filtrar por estado (PENDING, APPROVED, CANCELLED)", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas obtenida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Booking" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Reservas"],
        "summary": "Registrar una nueva reserva de motocicleta",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["motorcycleId", "startDate", "endDate"],
                "properties": {
                  "motorcycleId": { "type": "integer", "example": 1 },
                  "startDate": { "type": "string", "format": "date-time", "example": "2026-08-06T00:00:00Z" },
                  "endDate": { "type": "string", "format": "date-time", "example": "2026-08-10T00:00:00Z" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva creada exitosamente (estado PENDING)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Booking created successfully" },
                    "booking": { "$ref": "#/components/schemas/Booking" }
                  }
                }
              }
            }
          },
          "400": { "description": "Error en fechas, disponibilidad o conflicto de reservas solapadas" },
          "404": { "description": "Motocicleta no encontrada" }
        }
      }
    },
    "/api/bookings/{id}/status": {
      "put": {
        "tags": ["Reservas"],
        "summary": "Actualizar el estado de una reserva (Ej: Aprobar o Cancelar)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["PENDING", "APPROVED", "CANCELLED"], "example": "APPROVED" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Estado actualizado con éxito" },
          "400": { "description": "Estado inválido" },
          "403": { "description": "No autorizado (los usuarios comunes solo pueden cancelar)" },
          "404": { "description": "Reserva no encontrada" }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "tags": ["Reservas"],
        "summary": "Obtener detalles de una reserva específica",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la reserva obtenidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          },
          "403": { "description": "Acceso denegado (la reserva pertenece a otro usuario)" },
          "404": { "description": "Reserva no encontrada" }
        }
      },
      "delete": {
        "tags": ["Reservas"],
        "summary": "Eliminar el registro de una reserva",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Reserva eliminada con éxito" },
          "403": { "description": "Acceso denegado" },
          "404": { "description": "Reserva no encontrada" }
        }
      }
    },
...
  }
}